Refactored vue app into components
