--- static/js/whyis_vue/components/table-view/frontend/.eslintignore
@@ -1,4 +1,0 @@
-/build/
-/config/
-/dist/
-/*.js
