--- static/js/whyis_vue/components/table-view/frontend/config/dev.env.js
@@ -1,7 +1,0 @@
-'use strict'
-const merge = require('webpack-merge')
-const prodEnv = require('./prod.env')
-
-module.exports = merge(prodEnv, {
-  NODE_ENV: '"development"'
-})
