--- static/js/whyis_vue/components/table-view/frontend/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>apimock</title>
-    <link href="https://unpkg.com/tabulator-tables@4.1.4/dist/css/semantic-ui/tabulator_semantic-ui.min.css" rel="stylesheet">
-  </head>
-  <body>
-    <div id="app"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
