+++ static/js/whyis_vue/components/table-view/index.js
@@ -1,2 +1,3 @@
 // Matthew: import your module here
 // import "./your-module";
+import "./components/Tableview.vue";
