+++ static/package.json
@@ -5,6 +5,7 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
+    "tabulator-tables": "^4.4.3",
     "vue": "^2.6.10",
     "vue-material": "^1.0.0-beta-11",
     "webpack": "^4.35.2",
