+++ static/tabletest.html
@@ -1,0 +1,24 @@
+<html>
+    <head>
+        <link rel="stylesheet" href="https://unpkg.com/vue-material/dist/vue-material.min.css">
+        <link rel="stylesheet" href="https://unpkg.com/vue-material/dist/theme/default.css">
+        <link href="https://unpkg.com/tabulator-tables@4.1.4/dist/css/semantic-ui/tabulator_semantic-ui.min.css" rel="stylesheet">
+
+        <style>
+            #app {
+              font-family: 'Avenir', Helvetica, Arial, sans-serif;
+              -webkit-font-smoothing: antialiased;
+              -moz-osx-font-smoothing: grayscale;
+              text-align: center;
+              color: #2c3e50;
+              margin-top: 60px;
+            }
+        </style>
+    </head>
+    <body>
+        <div id="page">
+            <Tableview></Tableview>
+        </div>
+        <script src="js/whyis_vue_bundle.js"></script>
+    </body>
+</html>
